+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - Basic library function for parsing changelogs using `changelog-parser`
 - Determine next version numbers
+- Basic CLI to print the changelog to stdout